Refactor table creation dialog into its own component
